chore: Fixes and type annotations

+++ scripts/update_disco.py
@@ -16,11 +16,13 @@ import glob
 import json
 import logging
 import os
+import os.path
 import sys
 import urllib.request
+from typing import Any, Optional
 
 
-def main():
+def main() -> None:
     """The entrypoint for updating discovery documents
 
     This script pulls the latest discovery documents from the discovery service
@@ -28,8 +30,8 @@ def main():
     """
     logging.basicConfig(level=logging.INFO)
     os.chdir("discoveries")
-    index_document = load_index()
-    service_documents = load_documents(index_document)
+    index_document: DocumentInfo = load_index()
+    service_documents: list[DocumentInfo] = load_documents(index_document)
     delete_unused_files(index_document, service_documents)
     update_files(service_documents)
     update_index(index_document)
@@ -51,7 +53,13 @@ class DocumentInfo:
             and etag fields removed, for semantic comparisons
     """
 
-    def __init__(self, content, filename=None):
+    filename: str
+    content: bytes
+    json: Any
+    revision: Optional[str]
+    json_without_revision: Any
+
+    def __init__(self, content: bytes, filename: Optional[str] = None) -> None:
         """Analyzes document content
 
         Arguments:
@@ -74,66 +82,74 @@ class DocumentInfo:
             self.revision = None
 
 
-def load_index():
+def load_index() -> DocumentInfo:
     """Downloads the current discovery index and returns a document
 
     Returns:
         DocumentInfo -- Analysis of the discovery index content
     """
     print("LOADING index document ...")
-    with urllib.request.urlopen("https://discovery.googleapis.com/discovery/v1/apis") as f:
+    with urllib.request.urlopen(
+        "https://discovery.googleapis.com/discovery/v1/apis"
+    ) as f:
         if f.status != 200:
             sys.exit(f"Got HTTP status {f.status} for discovery index")
-        document = DocumentInfo(f.read())
-    if document.json == None:
+        document: DocumentInfo = DocumentInfo(f.read())
+    if document.json is None:
         sys.exit("Unable to parse discovery index")
     logging.info("Loaded index")
     return document
 
 
-def load_documents(index_document):
+def load_documents(index_document: DocumentInfo) -> list[DocumentInfo]:
     """Downloads all discovery documents in the index
 
     Arguments:
         index_document {DocumentInfo} -- The index
 
     Returns:
-        Sequence[DocumentInfo] -- List of discovery document content
+        list[DocumentInfo] -- List of discovery document content
     """
     print("LOADING service documents ...")
-    service_documents = []
+    service_documents: list[DocumentInfo] = []
     for item in index_document.json["items"]:
-        name = item["name"]
-        version = item["version"]
-        discovery_rest_url = item["discoveryRestUrl"]
-        filename = f"{name}.{version}.json"
+        name: str = item["name"]
+        version: str = item["version"]
+        discovery_rest_url: str = item["discoveryRestUrl"]
+        filename: str = f"{name}.{version}.json"
         # Sometimes the index lists services that don't exist. So log any
         # errors but don't let them crash the entire script.
         try:
             with urllib.request.urlopen(discovery_rest_url) as f:
                 if f.status != 200:
-                    logging.error(f"HTTP status {f.status} when loading {filename} from {discovery_rest_url}")
+                    logging.error(
+                        f"HTTP status {f.status} when loading {filename} from {discovery_rest_url}"
+                    )
                     continue
-                document = DocumentInfo(f.read(), filename)
+                document: DocumentInfo = DocumentInfo(f.read(), filename)
         except urllib.error.HTTPError:
             logging.error(f"Failed to load {filename} from {discovery_rest_url}")
             continue
         if document.revision == None:
-            logging.error(f"Malformed document for {filename} from {discovery_rest_url}")
+            logging.error(
+                f"Malformed document for {filename} from {discovery_rest_url}"
+            )
             continue
         logging.info(f"Loaded {filename} from {discovery_rest_url}")
         service_documents.append(document)
     return service_documents
 
 
-def delete_unused_files(index_document, service_documents):
+def delete_unused_files(
+    index_document: DocumentInfo, service_documents: list[DocumentInfo]
+) -> None:
     """Deletes files in the discoveries directory that are no longer used
 
     Arguments:
         index_document {DocumentInfo} -- The index
         service_documents {Sequence[DocumentInfo]} -- The discoveries
     """
-    expected_names = set([index_document.filename])
+    expected_names: set[str] = set([index_document.filename])
     for doc in service_documents:
         expected_names.add(doc.filename)
     for filename in glob.glob("*.json"):
@@ -143,38 +159,51 @@ def delete_unused_files(index_document, service_documents):
         os.remove(filename)
 
 
-def update_files(service_documents):
+def update_files(service_documents: list[DocumentInfo]) -> None:
     """Iterates the service discovery documents and updates the discoveries dir
 
     This updates only files where we downloaded a revision strictly greater
     than the existing revision, AND whose content has semantically changed.
 
     Arguments:
-        service_documents {Sequence[DocumentInfo]} -- The discoveries
+        service_documents {list[DocumentInfo]} -- The discoveries
     """
     for document in service_documents:
-        filename = document.filename
-        with open(filename, "rb+") as f:
-            existing = DocumentInfo(f.read(), filename)
-            if existing.revision != None:
-                if existing.revision > document.revision:
-                    logging.info(f"Existing revision {existing.revision} > updated revision {document.revision} for {filename}")
+        document_revision: Optional[str] = document.revision
+        assert document_revision is not None
+        filename: str = document.filename
+        if os.path.isfile(filename):
+            with open(filename, "rb+") as f:
+                existing: DocumentInfo = DocumentInfo(f.read(), filename)
+                existing_revision: str = existing.revision or "(unknown)"
+                if existing_revision > document_revision:
+                    logging.info(
+                        f"Existing revision {existing_revision} > updated revision {document_revision} for {filename}"
+                    )
                     continue
-                elif existing.revision == document.revision:
-                    logging.info(f"No change to revision {document.revision} for {filename}")
+                elif existing_revision == document_revision:
+                    logging.info(
+                        f"No change to revision {document_revision} for {filename}"
+                    )
                     continue
                 elif existing.json_without_revision == document.json_without_revision:
-                    logging.info(f"Revision updated {existing.revision} to {document.revision} but no other changes for {filename}")
+                    logging.info(
+                        f"Revision updated {existing_revision} to {document_revision} but no other changes for {filename}"
+                    )
                     continue
-                print(f"UPDATING revision {existing.revision} to {document.revision} for {filename}")
-            else:
-                print(f"WRITING new file {filename} at revision {document.revision}")
-            f.seek(0)
-            f.truncate()
-            f.write(document.content)
+                print(
+                    f"UPDATING revision {existing_revision} to {document_revision} for {filename}"
+                )
+                f.seek(0)
+                f.truncate()
+                f.write(document.content)
+        else:
+            print(f"WRITING new file {filename} at revision {document_revision}")
+            with open(filename, "wb") as f:
+                f.write(document.content)
 
 
-def update_index(index_document):
+def update_index(index_document: DocumentInfo) -> None:
     """Updates the index file in the discoveries dir
 
     Arguments:
